backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
anthropic>=0.40.0
requests==2.31.0
beautifulsoup4==4.12.3
youtube-transcript-api==0.6.2
python-dotenv==1.0.0
python-multipart==0.0.9
